sua giao dien kiem tra, them listening,listening type, quiztype,sua cac loi lat vat

+++ English_Quiz/Controllers/QuestionController.cs
@@ -13,14 +13,17 @@ namespace English_Quiz.Controllers
     public class QuestionController : BaseController
     {
         English_QuizEntities db = new English_QuizEntities();
+        [CheckPermission(PermissionName = "QuanLyCauHoi", Action = ConstantCommon.Action.View)]
         public ActionResult Index()
         {
             return View(db.Questions.ToList());
         }
+        [CheckPermission(PermissionName = "QuanLyCauHoi", Action = ConstantCommon.Action.Add)]
         public ActionResult CreateQuestion()
         {
-            ViewBag.ListLevel = new SelectList(db.Levels.ToList(), "LEVEL_ID", "LEVEL_NAME");
-            ViewBag.ListType = new SelectList(db.Types.ToList(), "TYPE_ID", "TYPE_NAME");
+           
+            ViewBag.ListType = new SelectList(db.Question_Type.ToList(), "TYPE_ID", "TYPE_NAME");
+            ViewBag.ListListening = new SelectList(db.Listenings.ToList(), "LISTENING_ID", "LISTENING_NAME_VN");
             return View();
         }
         [ValidateAntiForgeryToken]
@@ -37,11 +40,24 @@ namespace English_Quiz.Controllers
         }
         public ActionResult EditQuestion(string id)
         {
-            ViewBag.ListLevel = new SelectList(db.Levels.ToList(), "LEVEL_ID", "LEVEL_NAME");
-            ViewBag.ListType = new SelectList(db.Types.ToList(), "TYPE_ID", "TYPE_NAME");
+           
+            ViewBag.ListType = new SelectList(db.Question_Type.ToList(), "TYPE_ID", "TYPE_NAME");
+            ViewBag.ListListening = new SelectList(db.Listenings.ToList(), "LISTENING_ID", "LISTENING_NAME_VN");
             return View(db.Questions.FirstOrDefault(x=>x.QUESTION_ID == id));
         }
-        
+        [ValidateAntiForgeryToken]
+        [HttpPost]
+        public ActionResult EditQuestion(string id , Question question)
+        {
+            Question oldQuestion = db.Questions.FirstOrDefault(x => x.QUESTION_ID == id);
+            if (ModelState.IsValid)
+            {
+                db.Entry(oldQuestion).CurrentValues.SetValues(question);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            return null;
+        }
         public string GetAnswerByQuestionId()
         {
             string question_id = (Request["question_id"] == null) ? string.Empty : Request["question_id"].ToString();
@@ -61,7 +77,7 @@ namespace English_Quiz.Controllers
             return JsonConvert.SerializeObject(DTO);
         }
 
-        public void AddAnswer()
+        public string AddAnswer()
         {
             string description = (Request["description"] == null) ? string.Empty : Request["description"].ToString();
             string question_id = (Request["question_id"] == null) ? string.Empty : Request["question_id"].ToString();
@@ -74,6 +90,7 @@ namespace English_Quiz.Controllers
             ans.QUESTION_ID = question_id;
             db.Answers.Add(ans);
             db.SaveChanges();
+            return JsonConvert.SerializeObject(ans);
         }
         public void UpdateAnswer()
         {
@@ -83,14 +100,19 @@ namespace English_Quiz.Controllers
             int list_order = (Request["list_order"] == null) ? 0 : int.Parse(Request["list_order"].ToString());
             bool is_correct = (Request["is_correct"] == null) ? false : bool.Parse(Request["is_correct"].ToString());
             Answer ans = db.Answers.Where(x => x.ANSWER_ID == answer_id && x.QUESTION_ID == question_id).FirstOrDefault();
-            Answer nAns = new Answer();
-            nAns.DESCRIPTION = description;
-            nAns.IS_CORRECT = is_correct;
-            nAns.LIST_ORDER = list_order;
-            nAns.QUESTION_ID = question_id;
-            nAns.ANSWER_ID = answer_id;
-            db.Entry(ans).CurrentValues.SetValues(nAns);
-            db.SaveChanges();
+            if (ans != null)
+            {
+                Answer nAns = new Answer();
+                nAns.DESCRIPTION = description;
+                nAns.IS_CORRECT = is_correct;
+                nAns.LIST_ORDER = list_order;
+                nAns.QUESTION_ID = question_id;
+                nAns.ANSWER_ID = answer_id;
+
+                db.Entry(ans).CurrentValues.SetValues(nAns);
+                db.SaveChanges();
+            }
+            
         }
         public void DeleteAnswer()
         {
@@ -102,36 +124,50 @@ namespace English_Quiz.Controllers
         }
         public JsonResult DeleteQuestion(string id)
         {
-            try
+            int role = int.Parse(Session["Role"].ToString());
+            Permission permission = db.Permissions.FirstOrDefault(x => x.Role_Id == role);
+            if (permission.Is_Delete == true)
             {
-                var data = db.Questions.FirstOrDefault(x => x.QUESTION_ID == id);
-                var answerByQuestionId = db.Answers.Where(x => x.QUESTION_ID == id).ToList();
-                if (data == null)
+                try
+                {
+
+                    var data = db.Questions.FirstOrDefault(x => x.QUESTION_ID == id);
+                    var answerByQuestionId = db.Answers.Where(x => x.QUESTION_ID == id).ToList();
+                    if (data == null)
+                    {
+                        return Json(new
+                        {
+                            Success = false,
+                            Message = "Không tìm thấy đối tượng cần xóa."
+                        }, JsonRequestBehavior.AllowGet);
+                    }
+                    db.Questions.Remove(data);
+                    for (int i = 0; i < answerByQuestionId.Count; i++)
+                    {
+                        db.Answers.Remove(answerByQuestionId[i]);
+                    }
+                    db.SaveChanges();
+                    return Json(new
+                    {
+                        Success = true,
+                        Message = "Xóa thành công"
+                    }, JsonRequestBehavior.AllowGet);
+                }
+                catch (Exception e)
                 {
                     return Json(new
                     {
                         Success = false,
-                        Message = "Không tìm thấy đối tượng cần xóa."
+                        Message = "Không thể xóa đối tượng này. Vì sẽ ảnh hưởng đến dữ liệu khác." + e.Message
                     }, JsonRequestBehavior.AllowGet);
                 }
-                db.Questions.Remove(data);
-                for (int i = 0; i < answerByQuestionId.Count; i++)
-                {
-                    db.Answers.Remove(answerByQuestionId[i]);
-                }
-                db.SaveChanges();
-                return Json(new
-                {
-                    Success = true,
-                    Message = "Xóa thành công"
-                }, JsonRequestBehavior.AllowGet);
             }
-            catch (Exception e)
+            else
             {
                 return Json(new
                 {
                     Success = false,
-                    Message = "Không thể xóa đối tượng này. Vì sẽ ảnh hưởng đến dữ liệu khác." + e.Message
+                    Message = "Không có quyền xóa !"
                 }, JsonRequestBehavior.AllowGet);
             }
         }
